Implement data scraping for Roman villas
In the main program, the data scraping is performed from a given Wikipedia page using the newly implemented VillaDataScraper class. This class uses HtmlAgilityPack for HTML parsing, and includes methods for fetching and extracting villa data, along with saving this data into a JSON file. Additionally, new package references are added to the project file.